+++ qcparse/parsers/__init__.py
@@ -1,0 +1,2 @@
+# Required for parsers to register
+from .terachem import *  # noqa:  F403