+++ scripts/test.sh
@@ -1,3 +1,0 @@
-set -xe
-# Run tests
-poetry run pytest --cov-report html:htmlcov --cov